fix(MBA): int64_t loop variable in getDecisionVector — 64-bit refine soundness
Root cause of wrong 64-bit results on high-degree permutation cases:
getDecisionVector iterated the int64_t bit-cube result vector with an
'int' (32-bit) loop variable, silently truncating values >= 2^32.
The refine heuristic (tryRefine -> findTwoExpressionsByTwoValues ->
determineCombOfTwo) then built decision vectors from truncated values
and produced garbage candidates (e.g. 'c*0+d*0'); checkSolutionComplexity
compares metrics only and accepted them, replacing correct results.
At <=32-bit all values fit in int, which is why the bug only showed at
64-bit.

Minimal repro: 8991273086510694400*(x^y)-928395801376325632*(x|y)
simplified to 0 at 64-bit (not equivalent); correct at 32-bit.

Leftover of 1977f38, which fixed the same truncation class in the other
resultVector loops but missed this one.

+++ MBA/LinearSimplifier.cpp
@@ -372,7 +372,7 @@ std::vector<std::vector<Decision>> LinearSimplifier::getDecisionVector(
   const std::vector<int64_t> &v = (vec != nullptr) ? *vec : resultVector;
   std::vector<std::vector<Decision>> d;
 
-  for (int r : v) {
+  for (int64_t r : v) {
     std::vector<Decision> e;
     bool f = modInt(r - coeff1) == 0;
     bool s = modInt(r - coeff2) == 0;
